requirements.txt
Django==1.11.11
django-appconf==1.0.2
django-ckeditor==5.4.0
django-imagekit==4.0.2
django-qrcode==0.3
mysqlclient==1.3.4
olefile==0.45.1
pilkit==2.0
Pillow==4.2.1
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2018.3
six==1.11.0
